src/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
lxml==6.0.1
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.5
scapy==2.6.1
sniffio==1.3.1
soupsieve==2.7
speedtest-cli==2.1.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
